Move runtime state into "internal" header files.
