+++ Include/Python.h
@@ -133,4 +133,8 @@
 #include "fileutils.h"
 #include "pyfpe.h"
 
+#ifdef Py_BUILD_CORE
+#include "internal/_Python.h"
+#endif
+
 #endif /* !Py_PYTHON_H */
