+++ Include/internal/_ceval.h
@@ -1,0 +1,71 @@
+#ifndef _Py_CEVAL_H
+#define _Py_CEVAL_H
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#include "ceval.h"
+#include "compile.h"
+#include "pyatomic.h"
+
+#ifdef WITH_THREAD
+#include "pythread.h"
+#endif
+
+struct _pending_calls {
+    unsigned long main_thread;
+#ifdef WITH_THREAD
+    PyThread_type_lock lock;
+    /* Request for running pending calls. */
+    _Py_atomic_int calls_to_do;
+    /* Request for looking at the `async_exc` field of the current
+       thread state.
+       Guarded by the GIL. */
+    int async_exc;
+#define NPENDINGCALLS 32
+    struct {
+        int (*func)(void *);
+        void *arg;
+    } calls[NPENDINGCALLS];
+    int first;
+    int last;
+#else /* ! WITH_THREAD */
+    _Py_atomic_int calls_to_do;
+#define NPENDINGCALLS 32
+    struct {
+        int (*func)(void *);
+        void *arg;
+    } calls[NPENDINGCALLS];
+    volatile int first;
+    volatile int last;
+#endif /* WITH_THREAD */
+};
+
+#include "_gil.h"
+
+struct _ceval_globals {
+    int recursion_limit;
+    int check_recursion_limit;
+    /* Records whether tracing is on for any thread.  Counts the number
+       of threads for which tstate->c_tracefunc is non-NULL, so if the
+       value is 0, we know we don't have to check this thread's
+       c_tracefunc.  This speeds up the if statement in
+       PyEval_EvalFrameEx() after fast_next_opcode. */
+    int tracing_possible;
+    /* This single variable consolidates all requests to break out of
+       the fast path in the eval loop. */
+    _Py_atomic_int eval_breaker;
+#ifdef WITH_THREAD
+    /* Request for dropping the GIL */
+    _Py_atomic_int gil_drop_request;
+#endif
+    struct _pending_calls pending;
+    struct _gil_globals gil;
+};
+
+PyAPI_FUNC(void) _PyEval_Initialize(struct _ceval_globals *);
+
+#ifdef __cplusplus
+}
+#endif
+#endif /* !_Py_CEVAL_H */
